Fix bitwise operation between different enumeration types
On C++20 doing bitwise operation between different enumeration types
is deprecated.

+++ src/openrct2-ui/interface/Widget.cpp
@@ -570,7 +570,8 @@ static void WidgetCaptionDraw(rct_drawpixelinfo* dpi, rct_window& w, WidgetIndex
     }
     topLeft.x += width / 2;
     DrawTextEllipsised(
-        dpi, topLeft, width, widget->text, Formatter::Common(), { COLOUR_WHITE | COLOUR_FLAG_OUTLINE, TextAlignment::CENTRE });
+        dpi, topLeft, width, widget->text, Formatter::Common(),
+        { COLOUR_WHITE | static_cast<uint8_t>(COLOUR_FLAG_OUTLINE), TextAlignment::CENTRE });
 }
 
 /**
